refactor: move some helper functison to utiltities file
